+++ sdk/confidentialledger/Azure.Security.ConfidentialLedger/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Bugs Fixed
 
-- Allow some `HttpStatusCode.NotFound` occurrences in `PostLedgerEntryOperation` to account for unexpected loss of session stickiness. These errors may occur when the connected node changes and transactions have not been fully replicated.
+- Create a custom ResponseClassifier to allow `HttpStatusCode.NotFound` to be retriable. These errors may occur in the `PostLedgerEntryOperation` due to unexpected loss of session stickiness when the connected node changes and transactions have not been fully replicated.
 
 ## 1.2.0-beta.1 (2022-11-09)